+++ derive/tests/upload.rs
@@ -1,0 +1,4 @@
+mod schema_utils;
+mod upload {
+    mod upload_tests;
+}